Refactoring: remove useless parameter
This is part of request #10039: Retrieve FRS packages via REST

When we read the method FRSReleaseFactory::getFRSReleasesFromDb() we see that
the two optional parameters are strongly linked. You should:

* call the method with only the release id: getFRSReleasesFromDb($release_id)
* call the method with all parameters: getFRSReleasesFromDb($releases_id, $status_id, $group_id)

Each time the $group_id is used, this is when $status_id is given.

This is a necessary commit to be able to refactor and split this method like it
has been done for packages.

Change-Id: I56eee28dc92cec2323c82eb872e17eb3593ba9c7