Database + send to user
This is a part of story #9892: notify group of people in addition to individuals

Run forgeupgrade, hack your database and insert an entry in the table
tracker_global_notification_users. You need a notification_id, and
a user_id.

Then update an artifact. The given user should receive an email.

No changes from the user interface.

Change-Id: I17697be7624aea269bf89c4f7b5b7e2f9f5b9b30