Remove user from notifications
This is a part of story #9892: notify group of people in addition to individuals

Run forgeupgrade, hack your database and insert an entry in the table
tracker_global_notification_users. You need a notification_id, and
a user_id.

Then if you remove a notification all associated users are removed.

No changes from the user interface.

Change-Id: Ie3282b073d4459c94156e22262feb48d843a31db