Remove ugroup from notifications
This is a part of story #9892: notify group of people in addition to individuals

Run forgeupgrade, hack your database and insert an entry in the table
tracker_global_notification_ugroups. You need a notification_id, and
a ugroup_id.

Then if you remove a notification all associated ugroups are removed.

No changes from the user interface.

Change-Id: Ieae78c3d7e422661dd102030e6580028f58c1952