Deal with ugroup deletion
This is a part of story #9892: notify group of people in addition to individuals

Run forgeupgrade, hack your database and insert an entry in the table
tracker_global_notification_ugroups. You need a notification_id, and
a ugroup_id.

Then if remove the corresponding ugroup, you shouldn't see it anymore on notification.

Change-Id: If9c846c1bddc6b1942d4b24fbd33bc666672a24a