Docman notify user groups
This is part of story #9892: notify group of people in addition to
individuals

Run forgeupgrade, hack your database and insert an entry in the table
plugin_docman_notification_ugroups. You need an item_id, a ugroup_id and
a type ('plugin_docman').

Then update an item or a sub-item if you're monitoring a folder. Then
given ugroup members should receive an email.

No changes from the user interface.

Change-Id: I1e7175d0ae0803b85ab0b3965bc332dd0b0934de