Introduce edit mode in tracker notifications
Part of story #9892: notify group of people in addition to individuals

On tracker » admin » notifications, each global notification becomes
read only. User has to click on [Edit] link in order to be able to edit
the row. Two buttons appear to save or cancel the action.

When in edit mode, user cannot add ugroup/users yet. This will be done
in another commit.

The remove checkbox is not concerned by this edit mode.

Change-Id: I28919fa8306b3da1977ddeac9f9e13e82ef10957