Fix request #10084: Impossible to add a link to a User Story in a Task
Create a user story. Create a task.

Edit the task, change only the artifact link field: add the link to the
*parent* user story. Submit.

The changes should not be rejected. The link between the story and the
task is saved.

Remaining issue: if you add a link task -> story whereas there is
already a link story -> task, then the update is accepted, but you get
no feedback (and no changes).

Change-Id: Ibdbe91426ce25664cacdd26a9654ce73802549f4