Deal with item deletion
This is a part of story #9892: notify group of people in addition to
individuals

Run forgeupgrade, insert entries in the table
plugin_docman_notification_ugroups. You need an item_id and a ugroup_id.
You should use both types: 'plugin_docman', 'plugin_docman_cascade'

When you delete an item, all notifications on it should be removed.

Change-Id: I9f607d8d6ef523ac26d42a9f8808f25af4c302c9