Revert "REST route for submilestone creation must explicitly add the "_is_child" nature"
story #9857 see top backlog planning with mono-milestone

This reverts commit ab1851dc5797313190e87a9ba3e3fad84a71623f.

This commit introduce a REGRESSION
All new Backlogs element are added with _is_child nature.

The correct way to do this is:
Add a release in a release the _is_child nature is added
Add a backlog item it should be added with no nature link

We want to revert the commit, because we will
completly change the way it works.

Change-Id: Ie1ba9c8cc27e278bc916c8839f274b98dc14a472