Update signature of the Planning_MilestoneFactory class to follow changes in the Tuleap codebase
The commit 460f1e4cf311f7e68081268b3205309b4304c54e in the Tuleap repo has modified
the signature of the constructor of Planning_MilestoneFactory.

Without this changes a fatal error is generated when a SYSTEM_CHECK is processed and
standup summary are never sent.

This is part of story #9858: have general planning view managing mono-milestone

Change-Id: I85c487fb954f6c29f7ae11c1225f279f42bfefc8