request #10108: Cache user membership information
No need to check a second time in the db if user is member of a ugroup.

There may be a a lot of calls to the method isMemberOfUgroup when (we 
display an artifact in a project with lot of folders for example) and 
most of the time it has to perform the same query again and again. For 
example in a big setup isMemberOfUgroup is calling 106x the database :(

Change-Id: I1129b9e3ce3adea795812f6c8678a5e02306e596