request #9747: Add method to allow creation of project using REST API
One interesting thing that can be very useful is the integration between
the management team and development team. In normal organization, even
thought we can have a development team, there is the commercial team
that sales, the consultant or analyst. Most of this area uses ERPs to
perform all their managements with CRM and sales.

Change-Id: I20f73976d7291f2a5e4e3f4fef9447ac7e9e3b67