Detect SHA-1 collision based on the published attack [1]
It is a more generic way to detect a collision that the one we
currently have. The detection method we currently have is only
able to detect the published collision so far. With this
contribution we are able to detect collision created in a
similar way than the existing one. To have more information
on how the detection works please report to [2].

This is part of request #10015: SHA-1 colliding files can break Subversion repository

[1] https://shattered.io/static/shattered.pdf
[2] https://marc-stevens.nl/research/papers/C13-S.pdf

Change-Id: Id2b3e2bd800da8503850a70aa4852c022d317de9