Deal with user removed from private project
This is a part of story #9897 notify group in SVN

In a private project, add notifications for a user. Then remove this
user from the project. All her notifications should be removed.
In a public project, notifications should not be removed.

If only the user was in the notification (no additional email nor ugroup) then the
notification is completely removed.

Change-Id: I52755cae829b8e478a82be73a8a27fbe4f2db0f8