Force modal font-size to be the same as the body
Part of story #9839 Bootstrap 2 TLP theme

The modal content should have the same size as the main body content. If
the modal is included in a element that has a decreased size (eg: .9em),
then this should not affect the modal font-size.

Fix the modal that is shown when user clicks on the button [force cache
generation] for burndown field in an artifact (typically a sprint
burndown).

Change-Id: I2ee86e73043c1ffdd1f96a48bf0a121cef1dde9f