request #10156: Project imported from XML are no longer auto validated
The project is no longer autoactivated when there is services like FRS,
a fatal error is thrown because some directories are only created when
the project is activated.

This reverts commit 33b85bf67251d94fd727592592799580a8cdeec3. It means
that all the projects imported through XML will have the user doing
the import as the project's owner in the webhook payload.

Change-Id: I293cb15a60a96ab8b870a83fde6da6438d1be86e