request #10154: Collapse/Expand doesn't work on Docman Folders
This is introduced by I1a1191e6aa6474312ae03d22f17e6293f7de3a8d. The svg
added in the sidebar has attribute className that is not a regular
string but a SVGAnimatedString. Therefore the custom function
getElementsByClassNames failed.

Now that we have all decent browsers, we don't need anymore this custom
method and we can replace it by the standard querySelectorAll().

Remove also a debug method that were integrated when Firebug/Developer
console were not a thing yet.

Change-Id: I17e4abde3fe1f29570ea1e1ebf70c59b34fbae2d