Parse SVN logs after logrotation
This introduce a new RabbitMQ queue to manage parsing of logs.

Instead of keeping forever logs in /var/log/tuleap/year/month as it was done
on RHEL6, with RHEL7 only default logrotation is kept.

In addition to that, logs are parsed as they appear thanks to postrotate that
triggers a new event (via RabbitMQ) and consumed by a dedicated
worker (plugins/svn/bin/svn_stats_to_db.php).

The processing is not done directly in logrotate "postrotate" as the parsing
might be long.

Part of story #10167 have logs of svn activities done one backend-svn

Change-Id: Ieaa7df8f06c9c793e2cc8392fdf89393e1bc9e78