Project references override bugzilla reference
part of story #9770 cross references between Tuleap and Bugzilla

If you create a bugzilla server with the same keyword as a tracker in a
project, then in this project the references link toward the tracker, in
other projects the references link toward the bugzilla server.

Now when you create a Bugzilla server, then a corresponding system
reference is created. This reference cannot be removed nor "edited".

Users cannot create a reference with the nature 'bugzilla'.

Change-Id: I63659ed5b7cbe71b86fed801116276ec7943669b