request #10131: Images attached in PHPWiki can not be displayed in a wiki page when Internet Explorer 11 is used
The issue was caused because the value of the Content-Type header was set
surrounded with quotes (and thus invalid). Since commit
74889c9c59e498aab13dc9e25552d6387f725af6, IE11 does not do MIME Sniffing
anymore (which is good) and thus does not render the image. Recent or decent
browsers does not mime sniffing either but they know the web is deeply broken
so they do a bit of clean up before trying to use the content of a header.

This contribution only fix the symptom not the root cause, I'm afraid of
what's can break in PHPWiki if I do an attempt to properly fix it.

Change-Id: I9c998b98ee95521789c3195b8a6241e5add9812a