request #10179: Remove artifacts/:id/move from REST API
artifacts/:id/move is not a resource. We should not put action in the
name of a resource.

Furthermore, it is not at all a move; as said in epics #9563: "move =
new artifact + link + freeze". Therefore we cannot use a move "verb" on
/artifacts/:id resource, since it would imply to send a
"301 moved permanently" status code on a moved artifact, whereas the
resource still exist.

For now it is only a copy of an artifact, therefore one should use the
following route:

POST /artifacts
{
  tracker: {
    id: 109
  },
  from_artifact: {
    id: 42
  }
}

This will create a new artifact in the tracker 109, by copying the
values (actually the semantic title) of the artifact 42.

Change-Id: I46d21ed654bdc4444f2d7fe09619e88260e8a604