request #10227: Layout manager creates too many columns
The file LayoutManager.js is part of the assets *and* is included manually in
widgets.php. Since it is present twice in the page, two events are registered
to add a column and therefore two columns are added.

Now, one click to add a column or a row add only one column or one row.

Change-Id: I1b00967bda4fb0f506a740a036a8f8a3e2e50bd9