Refactoring: build sidebar in a dedicated object
Part of story #10173: Have multiple dashboards

In order to be able to include a project sidebar in BurningParrot, small
clean up must be done before in order to decrease the mess in the Layout
class.

Please note that the code has been moved and reorganized in best effort
mood. Code is not perfect, but cleaner than before. (I hope)

Things to check:
* admin entry is not visible for everybody
* plugin icon (like for git) is still displayed

Change-Id: If4a89dfb501280ae385b2d187e48e7e846a3636e