Display trove cat of project
Part of story #10173: Have multiple dashboards

If platform use trove cat, and project is categorized, then the
categories are displayed in project summary page header.

If trove cat is mandatory, and project is not categorized, then a
"empty state" is displayed. If current user is project admin, then a
feedback warning is displayed like in old summary page.

Change-Id: I10ceb7b163ada64b04f316c112df839a707caa8e