request #10202: Duplicate key error while importing Gitolite logs in the new format
The import processes the entries weeks by weeks but does not into account
that when a day is split in two, it is possible that Git accesses have
been done during the whole day.
This lead to a duplicate key error when you import a week because one part
of the day has already been imported into the new storage format but
we encounter the same day at a different moment.

Change-Id: I430a47ae08592cc554360b1aa517448b92071e18