request #10208 The description semantic can only use text fields
Given that this semantic is only used in the mail gateway context,
having a string field defined as the semantic descrption field will
not allow users to create an artifact by email.

For the old value defined with a string field,
a warning will be displayed in the semantic administration page
to warn the administrators about that point.

Change-Id: I8c34775855767208a843d3b30ec26b685c131b24