Display ImageViewer or MyProjects widget
This is a part of story #10173 Have multiple dashboards

How to test:

ImageViewer
===========

In FlamingParrot add a widget Image in a dashboard.

Update DashboardWidgetPresenter.php: owner_id, ownel_type and id of the
widget. For example if you created the widget in your personal page,
then :

  widget name = 'myimageviewer'
  owner_id    = your id
  owner_type  = 'u'
  widget id   = look into widget_image table

The widget should be displayed in the brand new dashboard (add lines and
cols in db for the dashboard before)

MyProjects
==========

Edit DashboardWidgetPresenter.php and uncomment 'myprojects' line. You
should see your projects, with basic tlp-styling (a full makeover of the
widget will be done later).

For the two widgets
===================

If widget has rss feed (MyProjects), then it is displayed as option in the dropdown.

If widget is editable (ImageViewer) then there is an option to edit it, currently it
does nothing.

There is an option to delete the widget, but it does nothing for now.

Change-Id: I2d391112ad76586a369932b49b6824f721699b55