User can add widget to dashboard
This is a part of story #10173 Have multiple dashboards

On new dashboards (my and projects), user can add widgets.

The widget is added at the top of the first column. If no column yet,
then the column is automatically created.

Unique widgets are not checked, it will be done in a dedicated
changeset.

Some widgets are broken (eg: Project contacts, Project classification,
…). They will be fixed in a dedicated changeset.

Change-Id: Icc91d92e055744e1664d03275aef12ab0e98d46f