request #10242: SSH keys validation accepts keys that are not valid
ssh-keygen has an undocummented feature (at least in the version used in CentOS6):
when it encounters a newline in a public key file it will extract the fingerprint
like it was an authorized_keys file. This allow users to upload public key with extra
parameters that have nothing to do in a public key file.

This issue has been introduced by commit 9f43853b3cf26c991c301f6fa60cec66f15d0dcb.

Change-Id: I87aa56e70dafc41303aaba3c27b79b8e69e17b79