BurningParrot project dropdowns has only one list of projects
request #10249 Projects dropdown should not distinguish projects that I admin and projects that I am member of

On a day to day basis, I don't need to have a separate list for projects that I
am member of and project that I admin. It is complicated to scan two separate
lists. It is not because I am admin of a project that I always do
administration on it.

Change-Id: I497298b441a15d5185f23ef15aadffa3bd592f0b