Display history in BP
This is part of story #10257: show artifact view history

If your history is empty, the dropdown still displays an empty state.
If it is not then there is an entry for each artifact that you visited.
the xref badge has the color of the tracker, and you can go to the
artifact, to the tracker or to the project from this entry.

Change-Id: I4971a57c482c25bfb1284d543bad69985070f369