Widgets have overflow-x: auto
This is part of story #10173 Have multiple dashboards

If you add a table renderer with lot of columns as a widget in a
dashboard, it should not screw up the whole page. instead the widget
respects the chosen layout and offers a scrollbar when needed.

Flex-basis is replaced by width else overflow-x property is not taken
into account by Chrome.

Change-Id: I23296967798015d0fd31839dc54599a575b3a9d9