request #10482: Improve performance of the page rendering engine
This contribution upgrades the Mustache library to v2.12

Recent versions of the library has improved the tokenizer
performance and allow to cache the templates.
These improvements yield a nice performance bonus:
 * in site admin on the system event administration page, I got a
~60% improvement of the execution time (678ms avg -> 272ms avg)
 * on a tracker report of Bug tracker with few artifacts, I got a
~9% improvement of the execution time (212ms avg -> 192ms avg)

Using a recent version will also allow us to use new features
of mustache.php, like blocks or helpers.

However this new version only supports the default behavior
of the mustache specification in case of a variable miss:
it silently returns an empty string.

Change-Id: I559c2c7d6024cffab20bfced348d404bc0c993c8