Main <a> should be a link not a div
This is part of story #10257: show artifact view history

In the history dropdown, I should be able to open the current entry in a new
tab (standard behavior of links). Since the current artifact is the main entry,
we focus on it. Unfortunately this won't work for extra links (quick links and
project name).

Change-Id: Iea1ca51bbdcc3c7190a03a7fb563a1cf309dcd4a