request #10319: Artifacts with a lot of attachments and a lot of comments take a long time to load
This contribution avoids to retrieve the same file informations
multiple times from the database.

To test, create an artifact, add a lot of attachments and comments.
By profiling the code you will see a lot of calls to Tracker_FileInfoDao::searchById,
this contribution reduces the number of needed calls to this method.

Change-Id: I42498e65ce868210551ce4edb09af3dc1674c7fd