Remove Widget::getPreferencesForm()
this is part of story #10173 Have multiple dashboards

This commit only removes unused getPreferencesForm from widget.
It should not break anything we should be able to add/read/update
any widget.

getPreferencesForm was only used for widget in legacy dashboards. As the latter
are gone we can safely remove it.

Change-Id: I306938cc59f3cf3fc491d5be5cf7c96f646c26fb