Use trash from fontawesome in MyProjects
This is a part of story #10173 Have multiple dashboards

First and foremost, I would like to apologize for the introduction of a
global deprecated function. I'm feeling very bad about this. However the
only purpose of this commit is to replace the ugly trash by the
fontawesome one.

Later, the widget will be rewamped (styling of the table, confirm()
replaced by a danger modal, …) and this function will be removed.

Change-Id: I287251191aa94516e6258f34a320e113dc353cf9