Remove orphan widgets after plugin removal
This is a part of story #10173 Have multiple dashboards

In a dashboard add a widget from a plugin (eg: project links). Uninstall
the plugin. The widget is removed from the db, and the dashboard is in a
clean state (eg: if it was the only widget, then the dashboard presents
an empty state).

Change-Id: I274bd0654b88548018c0a868b2fe58c5f0a4bfaf