request #10348: The iframe used to create artifact that must be added as artlink does not load
The fix done to prevent the bypass of the clickjacking protection for IE and Edge users
prevent to load any section of Tuleap in a frame (see commit 683ef92174a08cd).
This change break the modal on all browsers.
Since the modal can not be easily fixed the proposal is a compromise between
breaking the feature for all users and making all the pages of Tuleap vulnerable
to clickjacking for IE/Edge users. This contribution makes possible to frame the page
put in a iframe by the modal (and thus exposing to clickjacking this page for IE/Edge users).

Ultimately, the real fix should be to not use an iframe to create a modal.

Change-Id: I510135fbc92be52ee1a4aee3b84b69289e6d2a17