Create default dashboard at user creation
This is a part of story #10173 Have multiple dashboards

When a new account is created, a default layout is created with 2
default widgets (as of today): My projects, and My artifacts.

This remove the default My docman widget that was added in legacy
dashboards but not useful enough to be added now.

Change-Id: I0ad8090a71a68bf6f62619b4b2ed8b052dd984df