Respect column order of previous dashboards during migration
This is a part of story #10173 Have multiple dashboards

The previous dashboards were not using "ORDER BY id" to display columns of a
layout. Therefore we should not ORDER BY id during the migration.

This can be an issue for users that can see their widgets appearing in a
different order than before.

To reproduce, you can change the engine of layouts_rows_columns to MyISAM
and force the insertion of id in an odd order. For example:

INSERT INTO layouts_rows_columns (id, layout_row_id, width)
VALUES (8, 10, 33), (10, 10, 33), (9, 10, 33);

(Don't forget to add widgets in those columns)

With order by (previous version of this migration), columns are
displayed: 8, 9, 10.
Without order by (like it was displayed in legacy dashboards), columns
are displayed: 8, 10, 9.

Change-Id: Ia65a31fb2512141971ea260c2889616dc2a949e1