request #10380: Project members counters differ
On project dashboard, the number of members in the header should be the same as
the number displayed in the widget "Project Team". Event if the project has
suspended or deleted users as members. Furthermore, in this widget there is a
link to a really beautiful page that list members, the number of members
displayed should not differ either.

Change-Id: I378a2ffac376294544e243368a576415b04d7876