Display two different empty states
Part of story #10334: Widget Heartbeat - artifacts

If there are items, then nothing changes.
If there are no items at all (new project for example), then the empty
state says that there is no activity.
If all recent items cannot be displayed (because of permissions), then
the empty state reflects that.

Change-Id: I6ab0e81d119303d1d9df67ba2025f39444362f7b