Display git push in Heartbeat
Part of story #10336: Widget heartbeat - git push

Git pushes are now displayed in the heartbeat widget.

Personal repositories in the project are ignored.
If user cannot read the repository, then the push is not listed.
We skip the display of the user if it is unknown.

In another commit, the icon will be changed to git icon.

Change-Id: Ifcea7237091c6a4d24caab19f3775fbdb409969b