Introduce ChainOfResponsibility pattern to route Git requests
Part of request #10399: kill Git.class.php with fire

In order to split our famous god-object Git.class.php, a solution is to split
it in small objects. For example we can have a WebhookRouter that can handle
all request related to webhook management (and only them, SRP ftw).

By chaining small routers to handle git requests, we can reduce the complexity
of Git management.

This commit only deals with webhook management.

In ChainOfResponsibility pattern we chain the receiving objects and pass
the request along the chain until an object handles it. Don't worry, I
will not forget the fans of ascii diagrams:

     request = 【'edit-webhook', ...】

                    |
                    | process(request);
                    |
+----------------+ ‹´
|    A router    |    « Nope! »
+----------------+ \
        |           |
        |           | process(request);
        |           |
+-------ˇ--------+ ‹´
| Another router |    « Nope! »
+----------------+ \
        |           |
        |           | process(request);
        |           |
+-------ˇ--------+ ‹´
|      ...       |    « Nope! »
+----------------+ \
        |           |
        |           | process(request);
        |           |
+-------ˇ--------+ ‹´
| Webhook Router |    « Yay! I can deal with this request! \o/ »
+----------------+
        |
        |
        |
+-------ˇ--------+
|      ...       |
+----------------+
        |
        |
        |
+-------ˇ--------+
|   Legacy Git   | // This our god-object that will handle all remaining
+----------------+ // requests until we move them in dedicated objects

Change-Id: I131750e4778fa539e3ef053cb467a84b38614c87