request #10402: Use system CA bundle when verifying certificates for LDAPS communication in Python code
python-ldap library does not automatically choose a CA bundle
to do the verification of the chain of certificates so the connection always fail.
This contribution allow the library to use the same CA bundle used in PHP or
with most elements of the CentOS system. This CA bundle is provided through
the package ca-certificates.

Change-Id: Id09462f7fc75672336e7d3bae0e25de6543a7fd9