Save hook updates in project history
this is part of story #10214 manage svn "commit rules" with REST API

Please note that if you create a repository with settings by REST
route, you will have 2 entries in history.
In a few refacto commit we will make the distinguish between
POST creation with and without settings for route.

Change-Id: Ic6210bb5eaba52e3671385cfa2c061c5e3ab63ec