Inject last comments when viewing artifacts
Because we can fetch this in one query instead of fetching it bits by bits.
This saves 333 queries and reduce by 17% the wall time for my reference
artifact of 274 changesets.

This change the way artifacts are rendred for artifact view only ATM to reduce
the complexity, might be applied on other views later.

While I was running perf investigation I found a reference to:
request #9669 Field changes not displayed in last changeset comment
and gerrit #7069 but no evidence on what was the source of the issue.

Part of request #10069 Tracker comments don't scale up

Change-Id: Iccfc4c19242f34376c5357d508657e738f21ecde