Use angular v1.6 in Kanban
How to test:
- Test as much as possible all kanban actions
- Test as much as possible the angular artifact modal
- Check that realtime actions are still propagated

Notes:
- angular-ui-router was also upgraded
- radio buttons now perform a '===' comparison to find "checked" values,
which means we must use ng-value (value attribute means a string, not
integer value)
- $q will now throw an error for promises that are rejected but didn't
have an error handler. This is the main reason why unit-tests changed
- controllers now have a new built-in $onInit(). They no longer need to
call init() themselves.
- I don't think we should force all angular-artifact-modal dependent
apps to migrate to v1.6 right now. That is why, only in the modal, I did
not migrate all init() function calls to the new built-in $onInit().
Instead, kanban defines a $compileProvider config that lets us continue
as before, but is deprecated.

Change-Id: Ie9d8728f8ba724cd322a78ba0f7f3fd1006d694d