Fix request #10526: MOTD is not displayed in BurningParrot
Create file /etc/tuleap/site-content/en_US/others/motd.txt and add some
content in it. The motd is displayed both in FlamingParrot and
BurningParrot.

If the viewport is resized and the motd becomes bigger due to line
wrapping, then the main content is not hidden by the header.

On IE, in CompatibilityMode, a deprecation message is displayed. No
effort is made to properly style the motd (eg: top alignement) in
Compatibility Mode as per UXD request.

Background color is changed from bright blue to sidebar background
color.

Change-Id: Iffb47aa0b3a7bad0dab5f25d4dadf8249ee6e779