request #10527: Distributed Tuleap tests should timeout if they take too much time
Currently we wait until the platform is installed and the tests executed.
If you got an error we will retry to access the platform until it works.
In certain case it can never work (e.g. a PHP fatal error when accessing the page),
so we wait indefinitely.

We should not wait indefinitely, if the test seems to take too much time
to be executed we should collect logs and shut down.

Change-Id: Ica471e8b6ba007fd479b65a7e28c440424939dd9