Refactoring: DRY in Bind->getCriteria{From,Where}
This is part of request #10534: Wrong behavior when we search None in MSB

Move common code in the parent abstract class in order to avoid
duplication. The only difference is the management of ids to search with
a weird array_intersect combination for users/ugroups and a simple array_values
for static.

Change-Id: I9a210e26c44d7d3734fb93011b558cbfba5da490