request #10536: Clear caches tentative during the package upgrade should not show errors to the user
Clearing caches during the upgrade of the tuleap package is a tentative to limit issues
if the upgrade procedure is not followed perfectly (i.e. the tuleap service is not
restarted). Doing it at this point might raise warnings if there has been changes
if communication APIs between the core and the plugins has been modified since the
plugins packages are not yet updated. These warnings might mislead the users something
has gone wrong during the update. Since it is more of a supplementary fail safe
it seems reasonable to just hide it.

Change-Id: I7c3d8ce8a1d64c9d7b83d2cc2cd14376d19b0f67