Cache the information of a user being tracker administrator
This is a part of request #9981: Assigned/submitted Tracker widget is slow and unaccurate

When you use the My artifacts widget, it is needed to check if the user
can read or not the artifact. For that, it is verified if the user is tracker admin
which cost one SQL query.

With this contribution instead of executing one SQL query by artifact to check if the user
is tracker admin, we only execute one by tracker.

Change-Id: I75ce3e49e0cdb5ac788930c4e125b2080a3ff228