Fix build of the RPM package
With the release of RHEL7.4 a dependency of nodejs, http-parser, has been
moved to the base repository. As such it has been removed from EPEL.
CentOS 7.4 has not yet been released so nodejs can not be installed
on CentOS 7 at the moment since the depency can not be retrieved.

This contribution is a temporary fix until CentOS 7.4 is released,
http-parser is retrieved directly from the community build system.

This is part of request #10564: External plugins can not be built
due a missing dependency in the CentOS repository.

Change-Id: I221c24dd230277c938c3524dc56e4ea0e13eabac