Initiate Jenkinsfile
request #10569: The tests to be run should be dependent of the current
commit

Currently, when we add a new angular app in the sources of Tuleap, we
need to update the CI to add a specific job to run the corresponding
karma tests.

This is PITA since not all changesets under review can pass this new
job, which leads to rebase and frustration.

A proposal to fix this issue is to define the Jenkins pipeline directly
in the sources of Tuleap.

This is a first step that allows us to close the request. This commit
only add Jenkinsfile like it was defined on the pipeline
gerrit_pipeline_ut-php. Crazy groovy refactoring will be done in the
future.

You can see the result of this in the corresponding job:
https://ci.tuleap.org/jenkins/job/tuleap-gerrit-tests/
(Please note that this job is silent for now)

Change-Id: I6d96dab449a1db1a96841e018ea5d033f2a8cdc8