request #10394: User with no project must see the project dropdown in the navbar in BP
The project dropdown is now always present in a similar way than what is done in
the FP navbar.

In the Flaming Parrot navbar, the Projects dropdown is now hidden when the user
can not interact with it (i.e. not a member of a project, software map disabled and
project registration disabled)

Also, in the Burning Parrot navbar the button to go to the software map was always present
even if the software map was disabled in the local.inc.

Change-Id: I6b4303d831b691cd2e503b2f6cd8dad9c83086be