Unknown variables in Mustache templates throw an exception instead of being silently ignored.
Commit 351794e58b258435c146b81fe8d85472762905b9 has upgraded
the Mustache library but this new version silently ignores
unknown variable in templates and do not left the possibility
to configure this behaviour anymore.

This contribution is currently uses a dev branch in the
repository forked by Enalean. This branch will proposed as PR
upstream so please also review the branch in the forked repo.

To test, do not forget to clean your cache otherwise the compiled
templates will not take into account the new parameter.

This is part of request #10482: Improve performance of the page rendering engine

Change-Id: If21d4225ac3cf4e3309ae320b099725179ec302f