The notification sender is able to use a recent version of Mustache
A change in the Tuleap codebase has updated the version the mustache.php
library used (see commit 351794e58b258435c146b81fe8d85472762905b9 in the Tuleap
repository). To be able to use it the bot Mattermost plugins, the customization
done to manage Markdown templates needs to be adapted.

This contribution has a dependency on gerrit #9308.

This is part of request #10482: Improve performance of the page rendering engine

Change-Id: I6aea215d87d2a8e70344f0b24bcef3d699878c1b