request #10300: User lists when editing a card does not respect the preference of the user
When you change the assigned user through a card of the cardwall, the proposed
user lists should be ordered according the user's preference not the ID of the
proposed users.

A first contribution has attempted to fix the issue
(130c7d96a503f0656ee6b73c7387866187fefbfc) but while the users returned
by the JSON call were properly ordered the order were not kept at display.

Change-Id: I1dccf3aec6688cf533358b4504d342c5c1a7de8a