Display labels for a pull request
This is part of story #10498: Add labels on pull requests

.. IMPORTANT: this review has https://gerrit.tuleap.net/9316 as dependency

Run forge-upgrade.
Compile assets in core and in pull-requests (sass + angular).

In the database create 3 labels for a project that have pull requests
(in project_label table).
In the database link a pull requests to 2 project labels (in
plugin_pullrequest_label table).

Open the pull request in the browser, you should see the two chosen
labels under the sha1 of the pull request. (Those will be moved in a
dedicated commit).

If you switch to another pull request then the displayed labels change
accordingly.

Change-Id: I889c1b8880a413d43c7480b4e2fe1c2d12dd5104