request #10590: Propagate name change in shared fields
Create a shared field.
Update the label/description of the original field => changes are
propagated to target fields. (Works as expected)
Update the name of the original field => changes should be propagated to
target fields if possible. (A tracker cannot have to field with the same
name).

Example:

Tracker A: [summary, stag]
Tracker B: [summary, stag*]
Tracker C: [summary, stag*, status]

(*) the original field is `stag` in Tracker A

If I change the name in Tracker A to stage, then we will get:

Tracker A: [summary, stage]
Tracker B: [summary, stage*]
Tracker C: [summary, stage*, status]

Then if we change the name in Tracker A to status, then we will get:

Tracker A: [summary, status]
Tracker B: [summary, status*]
Tracker C: [summary, stage*, status]

Tracker C is not updated because it already has a status field.

Change-Id: I187587418be35b1eb91aa9a06e5bb8458400a5cb