Select the trackers
This is a part of story #10495 have base cross tracker search widget
(step 1)

You  can choose a project from all the projects you are a member of.
Then choose a tracker from this project, then add it to the list of
selected trackers for this widget.
The list of selected trackers will be saved in a next commit.
You can add up to 10 trackers only.
You can remove trackers from the list of selected trackers.

Change-Id: If57d256803de8aff2255b5a2514b2183dea1d3df