Slightly revamp pull request to display labels
This is part of story #10498: Add labels on pull requests

We now display the title of the current pull request on top of its
navigation bar. This allow us to display labels next to it. (Mock-up was
placing PR labels next to repo title, but it was confusing: user should
not think that labels refer to the repository.

We now have a better look and feel that mimic TLP panes for PR details
and discussion.

Change-Id: I78ae926f2691095bc173974cffb0560224622338