request #10597: Do not attempt to synchronize people when an error is encountered in one of the branches
The LDAP synchronization should not ignore when an error happens
when searching in one the branch that it not the last branch tested.
Currently only the status of the last tested branch is considered
which might lead to a lot of suspension if the users are not found
in the last branch and the search in the previous branches ended up
with errors.

Change-Id: I618144c5b051cac80dfc89d2f142311a20d43744