Edit labels of a pull request
This is part of story #10498: Add labels on pull requests

.. IMPORTANT: https://gerrit.tuleap.net/9351 is needed for this commit

I can edit¹ the pull request
============================

On a pull request without label:
- tags icon is displayed to add labels (with "Add label" label (!))

On a pull request with labels:
- labels are displayed with tag icon
-- tag icon becomes remove icon on mouseover
- plus icon is displayed to add labels (without "Add label" label)

Labels that are used in the project, in another pullrequest for example,
are provided in autocompletion.

Change (adding and removing) is automatically saved.

I cannot edit the pull request
==============================

On a pull request without label:
- Nothing is displayed

On a pull request with labels:
- labels are displayed with tag icon
-- tag icon does not change on mouseover
- there is no plus icon
- I cannot add or remove a tag

¹: ability to edit = user can merge.

Change-Id: Idf294664707ad728882a1c0855b265ca5d2c92a2